// backend/src/expedicao/enums/status-expedicao.enum.ts
export enum StatusExpedicao {
  AGUARDANDO_SEPARACAO = 'AGUARDANDO_SEPARACAO',
  PRONTO_PARA_RETIRADA = 'PRONTO_PARA_RETIRADA',
  EM_ROTA_DE_ENTREGA = 'EM_ROTA_DE_ENTREGA',
  AGUARDANDO_INSTALACAO = 'AGUARDANDO_INSTALACAO',
  ENTREGUE_FINALIZADO = 'ENTREGUE_FINALIZADO',
  ARQUIVADO = 'ARQUIVADO',
  DEVOLVIDA = 'DEVOLVIDA',
}

export const STATUS_EXPEDICAO_ATIVOS = new Set<string>([
  StatusExpedicao.AGUARDANDO_SEPARACAO,
  StatusExpedicao.PRONTO_PARA_RETIRADA,
  StatusExpedicao.EM_ROTA_DE_ENTREGA,
  StatusExpedicao.AGUARDANDO_INSTALACAO,
  StatusExpedicao.ENTREGUE_FINALIZADO,
  StatusExpedicao.ARQUIVADO,
]);